requirements.txt
python-telegram-bot==13.7
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
python-dotenv==0.19.0
web3==5.23.1
APScheduler==3.7.0
tronpy==0.2.5
solana==0.21.0
sui-sdk